moved classes + license headers
