+++ src/main/scala/nl/knaw/dans/easy/fedoratobag/FedoraProvider.scala
@@ -59,6 +59,11 @@ class FedoraProvider(fedoraClient: FedoraClient) {
       .flatMap(response => managed(response.getEntityInputStream))
   }
 
+  def datastream(objectId: String, streamId: String): ManagedResource[InputStream] = {
+    managed(FedoraClient.getDatastream(objectId, streamId).execute(fedoraClient))
+      .flatMap(response => managed(response.getEntityInputStream))
+  }
+
   def loadFoXml(objectId: String): Try[Elem] = {
     getObject(objectId).map(XML.load).tried
   }
