+++ src/main/scala/nl/knaw/dans/easy/fedoratobag/filter/FileFilterType.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.easy.fedoratobag
+package nl.knaw.dans.easy.fedoratobag.filter
 
 object FileFilterType extends Enumeration {
   type FileFilterType = Value
