+++ src/main/scala/nl/knaw/dans/easy/fedoratobag/versions/VersionInfo.scala
@@ -1,12 +1,24 @@
-package nl.knaw.dans.easy.fedoratobag.filter
+/**
+ * Copyright (C) 2020 DANS - Data Archiving and Networked Services (info@dans.knaw.nl)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package nl.knaw.dans.easy.fedoratobag.versions
 
-import java.time.DateTimeException
-
-import nl.knaw.dans.lib.string._
-import java.time.format.DateTimeFormatter
 import java.time.format.DateTimeFormatter.BASIC_ISO_DATE
 import java.time.{ DateTimeException, LocalDate }
 
+import nl.knaw.dans.lib.string._
 import org.joda.time.DateTime
 
 import scala.util.Try
@@ -39,6 +51,7 @@ object VersionInfo {
   }
 
   val easNameSpace = "http://easy.dans.knaw.nl/easy/easymetadata/eas/"
+
   private def isSelf(node: Node) = {
     val scheme = node.attribute(easNameSpace, "scheme")
       .map(_.text).getOrElse("")
